[MaterialDatePicker] Fix excessive whitespace
